Debounce scans for the same path
